requirements.txt
flask
flask-cors
torch
torchvision
numpy
opencv-python-headless
Pillow
scipy
gunicorn
